flask/requirements/dev.txt
-r prod.txt
black==23.9.1
factory-boy==3.3.0
httpx==0.25.0
isort==5.12.0
pytest-asyncio==0.21.1
pytest-factoryboy==2.5.1
pytest-sqlalchemy-mock==0.1.5
pytest==7.4.2
ruff==0.0.292
